api/requirements.txt
psycopg2-binary==2.8.5
fastapi==0.61.1
fastapi-utils==0.2.1
uvicorn==0.11.8
